Connecting the frontend and backendapi fix

+++ Backend/src/main/java/booking/repository/DriverRepository.java
@@ -3,4 +3,4 @@ package booking.repository;
 import booking.model.Driver;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DriverRepository extends JpaRepository<Driver, Long> {}
+public interface DriverRepository extends JpaRepository<Driver, Integer> {}
